refactor(auth): extract startAccountLogin to shared initiator module
把 popup/popupActions.js 的 startAccountLogin 搬到 scripts/auth/accountLoginInitiator.js，
為後續 onboarding tutorial 與 popup 共用同一登入觸發入口鋪路。

popupActions.js 改用 re-export 維持對外 API，既有 caller 與 4 個 popupActions.test.js
的相關測試完全不需動。

Plan: docs/plans/2026-05-18-onboarding-tutorial.md (Step 3 Commit 3a)
